% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspc.R
\name{fortify.SoilProfileCollection}
\alias{fortify.SoilProfileCollection}
\title{Fortify Method for SoilProfileCollections}
\usage{
\method{fortify}{SoilProfileCollection}(model, data, ...)
}
\arguments{
\item{model}{a SoilProfileCollection}

\item{data}{not used}

\item{...}{not used}
}
\value{
A data.frame
}
\description{
Called by \code{ggplot2::ggplot()}
}
