% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspc.R
\name{ggplot.SoilProfileCollection}
\alias{ggplot.SoilProfileCollection}
\title{ggplot S3 method for SoilProfileCollection objects}
\usage{
\method{ggplot}{SoilProfileCollection}(data, mapping = ggplot2::aes(), ..., environment = parent.frame())
}
\arguments{
\item{data}{A \code{SoilProfileCollection}}

\item{mapping}{Default \code{ggplot2::aes()}}

\item{...}{Additional arguments passed to \code{ggplot.default()} (not used)}

\item{environment}{Default: \code{parent.frame()}}
}
\value{
\code{gg} object from \code{ggplot2} package
}
\description{
ggplot S3 method for SoilProfileCollection objects
}
