% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspc-package.R
\docType{package}
\name{ggspc-package}
\alias{ggspc-package}
\alias{_PACKAGE}
\alias{ggspc.env}
\title{ggspc: Custom 'ggplot2' 'Stat', 'Geom', and 'theme' definitions for 'aqp' 'SoilProfileCollection' objects}
\description{
The goal of 'ggspc' is to provide custom 'Stat', 'Geom', and 'theme' definitions to enhance compatibility of 'SoilProfileCollection' objects with 'ggplot2'.
}
\author{
\strong{Maintainer}: Andrew Brown \email{andrew.g.brown@usda.gov}

}
\keyword{internal}
