% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_depth_weighted.R
\name{stat_depth_weighted}
\alias{stat_depth_weighted}
\title{\code{stat_depth_weighted()}: applies depth (thickness) weighted functions}
\usage{
stat_depth_weighted(
  data = NULL,
  mapping = NULL,
  geom = "point",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  FUN = stats::weighted.mean,
  from = 0,
  to = 200,
  ...
)
}
\arguments{
\item{data}{A \code{SoilProfileCollection}}

\item{mapping}{Aesthetic mapping constructed via \code{ggplot2::aes()}}

\item{geom}{Default: "point"}

\item{position}{Default: \code{"identity"}}

\item{na.rm}{Default: \code{FALSE}}

\item{show.legend}{Default: \code{NA}}

\item{inherit.aes}{Default: \code{TRUE}}

\item{FUN}{A function. First argument is values, second argument is weights. Should accept \code{na.rm} argument. Default: \code{weighted.mean()}}

\item{from}{Top depth of weighted function interval. Default: \code{0}.}

\item{to}{Bottom depth of weighted function interval. Default: \code{200}.}

\item{...}{additional arguments passed as parameters to \code{ggplot2::layer()}}
}
\value{
A \code{ggplot2::layer()} (a combination of data, Stat and \code{geom} with a potential position adjustment)
}
\description{
A \code{ggplot2}-style layer using the \code{StatDepthWeighted} statistic
}
