% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_pscs_wtd.R
\name{stat_pscs_wtd}
\alias{stat_pscs_wtd}
\title{\code{stat_pscs_wtd()}}
\usage{
stat_pscs_wtd(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{a SoilProfileCollection}

\item{geom}{default: point}

\item{position}{default: identity}

\item{na.rm}{default: false}

\item{show.legend}{default: NA}

\item{inherit.aes}{default: TRUE}

\item{...}{additonal arguments passed as parameters to \code{ggplot2::layer()}}
}
\value{
A \code{ggplot2::layer()} (a combination of data, Stat and \code{geom} with a potential position adjustment)
}
\description{
A \code{ggplot2}-style layer using the \code{StatPSCSWeightedAverage} statistic
}
