% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspc-package.R
\docType{package}
\name{ggspc-package}
\alias{ggspc}
\alias{ggspc-package}
\title{ggspc: Custom 'ggplot2' 'Stat', 'Geom', and 'theme' definitions for 'aqp' 'SoilProfileCollection' objects}
\description{
Provides custom 'Stat', 'Geom' and 'theme' definitions for 'SoilProfileCollection' objects compatibility with 'ggplot2' via 'ggproto'
}
\author{
\strong{Maintainer}: Andrew Brown \email{andrew.g.brown@usda.gov}

}
\keyword{internal}
